--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -22,6 +22,7 @@ Imports:
     pROC,
     RcppEigen,
     rstan,
+    SLmetrics,
     stats
 URL: https://github.com/swarnendu-stat/bnns, https://swarnendu-stat.github.io/bnns/
 BugReports: https://github.com/swarnendu-stat/bnns/issues