--- a/R/measure.R
+++ b/R/measure.R
@@ -27,8 +27,8 @@ measure_cont <- function(obs, pred) {
   error <- (obs - pred)
   error <- error[is.finite(error)]
   return(list(
-    rmse = sqrt(mean(error^2)),
-    mae = mean(abs(error))
+    rmse = SLmetrics::rmse(actual = obs, predicted = pred),
+    mae  = SLmetrics::mae(actual = obs, predicted = pred)
   ))
 }