Hi @swarnendu-stat 

Awesome R package, looks promising. I wanted to contribute with my own R package [{SLmetrics}](https://github.com/serkor1/SLmetrics) that is focused on performance evaluation of AI/ML models, and thought that your package was good fit as it is still in the early phase.

Would you be interested in a Dependency/Suggests on  [{SLmetrics}](https://github.com/serkor1/SLmetrics)? I will happily contribute and help out in the initial phase of the addition, and maintain it upstream. If not, feel free to close the PR! :smile: 

Below is the PR: 

## :books: What?

This PR adds [{SLmetrics}](https://github.com/serkor1/SLmetrics) to the Suggests-section of the Description file, and implements its functionality in `measure_cont()`. It is an entirely experimental and suggestive addition.[^1]

### :eyes: Why?

[{SLmetrics}](https://github.com/serkor1/SLmetrics) is built on C++ and primarily uses pointer arithmetic to calculate performance measures - so it reduces memory consumption, and optimizes the speed. See the benchmark below:

``` r
obs  <- rnorm(n = 1e6)
pred <- obs + rnorm(n = 1e6)

rmse <- function(obs, pred) {
  error <- (obs - pred)
  error <- error[is.finite(error)]
  sqrt(mean(error^2))
}


bench::mark(
  `{Original}`  = rmse(obs, pred),
  `{SLmetrics}` = SLmetrics::rmse(
    actual = obs,
    predicted = pred
  ),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 × 6
#>   expression       min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>  <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 {Original}       7ms   7.41ms      91.3   34.35MB     102.
#> 2 {SLmetrics}    657µs 662.37µs    1491.     5.91MB       0
```

<sup>Created on 2025-03-24 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>

So there is some RAM to be salvaged during the training, and evaluation, of the Neural Network..


[^1]: I will implement the remaining measures if you think its a good idea to add this dependency.